// src/app/shared/models/shift-user.model.ts
export class ShiftHandoverModel {
    Id: number;
    RequestId: string;
    LogbookId: string;
    PresentSituation: string;
    LoadCutDetails: string;
    DateTelemetryDetails: string;
    TransmissionSystemDetails: string;
    IsOutageTrippingRecorded: string;
    NewTransmissionOutageDetails: string;
    ImportantInformation: string;
    ImportantEvents: string;
}